Cardano.Node.Socket.Emulator.Mock

Documentation

healthcheck :: Monad m => m NoContent

consumeEventHistory :: MonadIO m => MVar AppState -> m EmulatorLogs

slotCoordinator :: SlotConfig -> ServerHandler -> IO a

Calls addBlock at the start of every slot, causing pending transactions to be validated and added to the chain.