Cardano.Node.Socket.Emulator.Server

Documentation

data ServerHandler

A handler used to pass around the path to the server and channels used for controlling the server.

runServerNode :: MonadIO m => Trace IO EmulatorMsg -> FilePath -> Integer -> AppState -> Params -> m ServerHandler

Start the server in a new thread, and return a server handler used to control the server

processBlock :: MonadIO m => ServerHandler -> m Block

modifySlot :: MonadIO m => (Slot -> Slot) -> ServerHandler -> m Slot

addTx :: MonadIO m => ServerHandler -> Tx BabbageEra -> m ()

processChainEffects :: Trace IO EmulatorMsg -> Params -> MVar AppState -> EmulatorT IO a -> IO a