Cardano.Node.Socket.Emulator.Types

Description

This module exports data types for logging, events and configuration

Documentation

type Tip = Tip (CardanoBlock StandardCrypto)

type TxPool = [CardanoTx]

data SocketEmulatorState

Constructors

SocketEmulatorState

Fields

_emulatorState :: EmulatorState

_channel :: TChan Block

_tip :: Tip

tip :: Lens' SocketEmulatorState Tip

emulatorState :: Lens' SocketEmulatorState EmulatorState

channel :: Lens' SocketEmulatorState (TChan Block)

data NodeServerConfig

Node server configuration

Constructors

NodeServerConfig

Fields

nscBaseUrl :: BaseUrl

base url of the service

nscInitialTxWallets :: [WalletNumber]

The wallets that receive money from the initial transaction.

nscSocketPath :: FilePath

Path to the socket used to communicate with the server.

nscKeptBlocks :: Integer

The number of blocks to keep for replaying to newly connected clients

nscSlotConfig :: SlotConfig

Beginning of slot 0.

nscNetworkId :: NetworkId

NetworkId that's used with the CardanoAPI.

nscProtocolParametersJsonPath :: Maybe FilePath

Path to a JSON file containing the protocol parameters

defaultNodeServerConfig :: NodeServerConfig

data AppState

Application State

Constructors

AppState

Fields

_socketEmulatorState :: SocketEmulatorState

blockchain state

_emulatorLogs :: EmulatorLogs

history of all log messages

socketEmulatorState :: Lens' AppState SocketEmulatorState

emulatorLogs :: Lens' AppState EmulatorLogs

fromEmulatorChainState :: MonadIO m => EmulatorState -> m SocketEmulatorState

initialChainState :: MonadIO m => Map CardanoAddress Value -> m SocketEmulatorState

ChainState with initial values

getChannel :: MonadIO m => MVar AppState -> m (TChan Block)

getTip :: MonadIO m => MVar AppState -> m Tip

setTip :: MonadIO m => MVar AppState -> Block -> m ()

data CNSEServerLogMsg

Top-level logging data type for structural logging inside the CNSE server.

Constructors

StartingSlotCoordination UTCTime Millisecond

StartingCNSEServer Int

ProcessingEmulatorMsg EmulatorMsg

newtype BlockId

The node protocols require a block header type.

Constructors

BlockId

Fields

getBlockId :: ShortByteString

blockId :: Block -> BlockId

A hash of the block's contents.

nodeToClientVersion :: NodeToClientVersion

Protocol versions

nodeToClientVersionData :: NodeToClientVersionData

A temporary definition of the protocol version. This will be moved as an argument to the client connection function in a future PR (the network magic number matches the one in the test net created by scripts)

doNothingInitiatorProtocol :: MonadTimer m => RunMiniProtocol 'InitiatorMode ByteString m a Void

A protocol client that will never leave the initial state.

doNothingResponderProtocol :: MonadTimer m => RunMiniProtocol 'ResponderMode ByteString m Void a

epochSlots :: EpochSlots

Boilerplate codecs used for protocol serialisation.

The number of epochSlots is specific to each blockchain instance. This value is what the cardano main and testnet uses. Only applies to the Byron era.

codecVersion :: BlockNodeToClientVersion (CardanoBlock StandardCrypto)

codecConfig :: CodecConfig (CardanoBlock StandardCrypto)

nodeToClientCodecs :: forall m. MonadST m => ClientCodecs (CardanoBlock StandardCrypto) m

chainSyncCodec :: block ~ CardanoBlock StandardCrypto => Codec (ChainSync block (Point block) Tip) DeserialiseFailure IO ByteString

These codecs are currently used in the mock nodes and will probably soon get removed as the mock nodes are phased out.

txSubmissionCodec :: block ~ CardanoBlock StandardCrypto => Codec (LocalTxSubmission (GenTx block) (ApplyTxErr block)) DeserialiseFailure IO ByteString

toCardanoBlock :: Header StandardCrypto -> Block -> CardanoBlock StandardCrypto

fromCardanoBlock :: CardanoBlock StandardCrypto -> Block